+++ build-bleeding-edge-toolchain.sh
@@ -3,7 +3,7 @@
 #
 # file: build-bleeding-edge-toolchain.sh
 #
-# author: Copyright (C) 2016-2017 Freddie Chopin http://www.freddiechopin.info http://www.distortec.com
+# author: Copyright (C) 2016-2018 Freddie Chopin http://www.freddiechopin.info http://www.distortec.com
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
 # distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -12,7 +12,7 @@
 set -e
 set -u
 
-binutilsVersion="2.29.1"
+binutilsVersion="2.30"
 expatVersion="2.2.5"
 gccVersion="7.3.0"
 gdbVersion="8.0.1"